Alterações finais da viewmodel

+++ CarCRM/Controllers/PerfilsController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using CarCRM.Models;
 using CarCRM.Data;
+using CarCRM.ViewModels;
 
 public class PerfilsController : Controller
 {
@@ -16,7 +17,13 @@ public class PerfilsController : Controller
     // GET: PERFILS
     public async Task<IActionResult> Index()    
     {
-        return View(await _context.Perfis.ToListAsync());
+        var perfis = await _context.Perfis.ToListAsync();
+        var perfilViewModel = perfis.Select(p => new PerfilViewModel
+        {   
+            Id = p.Id,
+            Nome = p.Nome
+        }).ToList();
+        return View(perfilViewModel);
     }
 
     // GET: PERFILS/Details/5
@@ -34,7 +41,13 @@ public class PerfilsController : Controller
             return NotFound();
         }
 
-        return View(perfil);
+        var PerfilViewModel = new PerfilViewModel
+        {
+            Id = perfil.Id,
+            Nome = perfil.Nome
+        };
+
+        return View(PerfilViewModel);
     }
 
     // GET: PERFILS/Create
@@ -48,15 +61,20 @@ public class PerfilsController : Controller
     // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Create([Bind("Id,Nome")] Perfil perfil)
+    public async Task<IActionResult> Create(PerfilViewModel perfilViewModel)
     {
         if (ModelState.IsValid)
         {
+            var perfil = new Perfil
+            {
+                Nome = perfilViewModel.Nome,
+            };
+
             _context.Add(perfil);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
-        return View(perfil);
+        return View(perfilViewModel);
     }
 
     // GET: PERFILS/Edit/5
@@ -72,7 +90,13 @@ public class PerfilsController : Controller
         {
             return NotFound();
         }
-        return View(perfil);
+
+        var perfilViewModel = new PerfilViewModel
+        {
+            Id = perfil.Id,
+            Nome = perfil.Nome
+        };
+        return View(perfilViewModel);
     }
 
     // POST: PERFILS/Edit/5
@@ -80,9 +104,9 @@ public class PerfilsController : Controller
     // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Edit(int? id, [Bind("Id,Nome")] Perfil perfil)
+    public async Task<IActionResult> Edit(int? id, PerfilViewModel perfilViewModel)
     {
-        if (id != perfil.Id)
+        if (id != perfilViewModel.Id)
         {
             return NotFound();
         }
@@ -91,12 +115,18 @@ public class PerfilsController : Controller
         {
             try
             {
+                var perfil = _context.Perfis.Find(id);
+                if (perfil == null)
+                {
+                    return NotFound();
+                }
+                perfil.Nome = perfilViewModel.Nome;
                 _context.Update(perfil);
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException)
             {
-                if (!PerfilExists(perfil.Id))
+                if (!PerfilExists(perfilViewModel.Id))
                 {
                     return NotFound();
                 }
@@ -107,7 +137,7 @@ public class PerfilsController : Controller
             }
             return RedirectToAction(nameof(Index));
         }
-        return View(perfil);
+        return View(perfilViewModel);
     }
 
     // GET: PERFILS/Delete/5
@@ -125,7 +155,13 @@ public class PerfilsController : Controller
             return NotFound();
         }
 
-        return View(perfil);
+        var perfilViewModel = new PerfilViewModel
+        {
+            Id = perfil.Id,
+            Nome = perfil.Nome
+        };
+
+        return View(perfilViewModel);
     }
 
     // POST: PERFILS/Delete/5
@@ -134,6 +170,13 @@ public class PerfilsController : Controller
     public async Task<IActionResult> DeleteConfirmed(int? id)
     {
         var perfil = await _context.Perfis.FindAsync(id);
+        bool hasUsuarios = await _context.Usuarios.AnyAsync(u => u.PerfilId == id);
+        if (hasUsuarios)
+        {
+            TempData["Erro"] = "Não é possível excluir este perfil porque há usuários vinculados a ele.";
+            return RedirectToAction(nameof(Index));
+        }
+
         if (perfil != null)
         {
             _context.Perfis.Remove(perfil);
